created class for file reading; eliminated TimeTT module; deleted module headerRelated; changed the class Time name to TimeTT; modified all the constructors of all classes and updated tukTuk to suport all the previous alterations
