+++ Driver.py
@@ -5,7 +5,7 @@
 # 43134 Luís Filipe Leal Campos
 # 48392 Mariana Vieira De Almeida Nave
 
-from Time import Time
+from TimeTT import Time
 
 
 class Driver(object):
