--- constants.py
@@ -1,23 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# 2015-2016 Complementos de Programacao
-# Grupo 3
-# 43134 Luís Filipe Leal Campos
-# 48392 Mariana Vieira De Almeida Nave
-
-# This module records the constants used in the application
-
-# In a file:
-# Number of line in a header
-NUMBEROfLinesInHeader = 7
-
-# In a the header list:
-# Index of elements with name of company
-INDEXCompany = 1
-
-# Index of elements with the date
-INDEXDate = 3
-
-# Index of elements with the period
-INDEXPeriod = 5
-
